+++ Sources/KeyboardKit/Keyboard/KeyboardCasing.swift
@@ -10,10 +10,6 @@ import Foundation
 
 /**
  This enum lists the various shift states a keyboard can use.
- 
- Note that `neutral` will be removed in KeyboardKit 5, but I
- have chosen to not mark it as such, since that would create
- a bunch of warnings within the framework.
  */
 public enum KeyboardCasing: String, Codable, Identifiable {