+++ Sources/KeyboardKit/Layout/KeyboardLayoutItemSize.swift
@@ -29,10 +29,10 @@ public struct KeyboardLayoutItemSize: Equatable {
     /**
      The declarative width of the item.
      */
-    public let width: KeyboardLayoutItemWidth
+    public var width: KeyboardLayoutItemWidth
     
     /**
      The fixed height of the item.
      */
-    public let height: CGFloat
+    public var height: CGFloat
 }